Move items_identified to avatar
And deal with all the issues that cropped up
 - Make several functions in player pure virtual
 - Make newcharacter.cpp use avatar
 - Temporarily a few nasty dynamic_casts until migration to avatar
   is finished.
