+++ src/activity_handlers.cpp
@@ -2639,7 +2639,11 @@ void activity_handlers::butcher_do_turn( player_activity *, player *p )
 
 void activity_handlers::read_finish( player_activity *act, player *p )
 {
-    p->do_read( *act->targets.front().get_item() );
+    if( avatar *u = dynamic_cast<avatar *>( p ) ) {
+        u->do_read( *act->targets.front().get_item() );
+    } else {
+        act->set_to_null();
+    }
     if( !act ) {
         p->add_msg_if_player( m_info, _( "You finish reading." ) );
     }
