+++ src/game_inventory.cpp
@@ -895,12 +895,19 @@ class read_inventory_preset: public pickup_inventory_preset
                     return unknown;
                 }
                 std::vector<std::string> dummy;
-                const player *reader = p.get_book_reader( *loc, dummy );
+
+                // This is terrible and needs to be removed asap when this entire file is refactored
+                // to use the new avatar class
+                const avatar *u = dynamic_cast<const avatar *>( &p );
+                if( !u ) {
+                    return std::string();
+                }
+                const player *reader = u->get_book_reader( *loc, dummy );
                 if( reader == nullptr ) {
                     return std::string();  // Just to make sure
                 }
                 // Actual reading time (in turns). Can be penalized.
-                const int actual_turns = p.time_to_read( *loc, *reader ) / to_moves<int>( 1_turns );
+                const int actual_turns = u->time_to_read( *loc, *reader ) / to_moves<int>( 1_turns );
                 // Theoretical reading time (in turns) based on the reader speed. Free of penalties.
                 const int normal_turns = get_book( loc ).time * reader->read_speed() / to_moves<int>( 1_turns );
                 const std::string duration = to_string_approx( time_duration::from_turns( actual_turns ), false );
@@ -918,8 +925,15 @@ class read_inventory_preset: public pickup_inventory_preset
         }
 
         std::string get_denial( const item_location &loc ) const override {
+            // This is terrible and needs to be removed asap when this entire file is refactored
+            // to use the new avatar class
+            const avatar *u = dynamic_cast<const avatar *>( &p );
+            if( !u ) {
+                return std::string();
+            }
+
             std::vector<std::string> denials;
-            if( p.get_book_reader( *loc, denials ) == nullptr && !denials.empty() ) {
+            if( u->get_book_reader( *loc, denials ) == nullptr && !denials.empty() ) {
                 return denials.front();
             }
             return pickup_inventory_preset::get_denial( loc );
@@ -953,7 +967,12 @@ class read_inventory_preset: public pickup_inventory_preset
         }
 
         bool is_known( const item_location &loc ) const {
-            return p.has_identified( loc->typeId() );
+            // This is terrible and needs to be removed asap when this entire file is refactored
+            // to use the new avatar class
+            if( const avatar *u = dynamic_cast<const avatar *>( &p ) ) {
+                return u->has_identified( loc->typeId() );
+            }
+            return false;
         }
 
         int get_known_recipes( const islot_book &book ) const {
