+++ src/savegame_json.cpp
@@ -535,21 +535,6 @@ void Character::store( JsonOut &json ) const
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 ///// player.h, avatar + npc
 
-/*
- * Prepare a json object for player, including player specific data, and data common to
- * players and npcs.
- * TODO: Make player abstract and delete this
- */
-void player::serialize( JsonOut &json ) const
-{
-    json.start_object();
-    // This must be after the json object has been started, so any super class
-    // puts their data into the same json object.
-    store( json );
-
-    json.end_object();
-}
-
 /**
  * Gather variables for saving. These variables are common to both the avatar and npcs.
  */
@@ -656,17 +641,6 @@ void player::store( JsonOut &json ) const
         }
         json.end_array();
     }
-}
-
-/*
- * Load player (soon to be renamed to survivor) from ginormous json blob.
- * Used for avatars and npcs.
- * TODO: Make player abstract and delete this
- */
-void player::deserialize( JsonIn &jsin )
-{
-    JsonObject data = jsin.get_object();
-    load( data );
 }
 
 /**
