+++ tests/crafting_test.cpp
@@ -117,7 +117,7 @@ TEST_CASE( "recipe_subset" )
 TEST_CASE( "available_recipes", "[recipes]" )
 {
     const recipe *r = &recipe_id( "brew_mead" ).obj();
-    player dummy;
+    avatar dummy;
 
     REQUIRE( dummy.get_skill_level( r->skill_used ) == 0 );
     REQUIRE_FALSE( dummy.knows_recipe( r ) );
@@ -212,7 +212,7 @@ TEST_CASE( "available_recipes", "[recipes]" )
 TEST_CASE( "crafting_with_a_companion", "[.]" )
 {
     const recipe *r = &recipe_id( "brew_mead" ).obj();
-    player dummy;
+    avatar dummy;
 
     REQUIRE( dummy.get_skill_level( r->skill_used ) == 0 );
     REQUIRE_FALSE( dummy.knows_recipe( r ) );
