+++ tests/reload_option.cpp
@@ -4,11 +4,11 @@
 #include "catch/catch.hpp"
 #include "item.h"
 #include "item_location.h"
-#include "player.h"
+#include "avatar.h"
 
 TEST_CASE( "revolver_reload_option", "[reload],[reload_option],[gun]" )
 {
-    player dummy;
+    avatar dummy;
 
     item &gun = dummy.i_add( item( "sw_619", 0, 0 ) );
     item &ammo = dummy.i_add( item( "38_special", 0, gun.ammo_capacity() ) );
@@ -36,7 +36,7 @@ TEST_CASE( "revolver_reload_option", "[reload],[reload_option],[gun]" )
 
 TEST_CASE( "magazine_reload_option", "[reload],[reload_option],[gun]" )
 {
-    player dummy;
+    avatar dummy;
 
     item &magazine = dummy.i_add( item( "glockmag", 0, 0 ) );
     item &ammo = dummy.i_add( item( "9mm", 0, magazine.ammo_capacity() ) );
@@ -55,7 +55,7 @@ TEST_CASE( "magazine_reload_option", "[reload],[reload_option],[gun]" )
 
 TEST_CASE( "belt_reload_option", "[reload],[reload_option],[gun]" )
 {
-    player dummy;
+    avatar dummy;
 
     item &belt = dummy.i_add( item( "belt308", 0, 0 ) );
     item &ammo = dummy.i_add( item( "308", 0, belt.ammo_capacity() ) );
@@ -79,7 +79,7 @@ TEST_CASE( "belt_reload_option", "[reload],[reload_option],[gun]" )
 
 TEST_CASE( "canteen_reload_option", "[reload],[reload_option],[liquid]" )
 {
-    player dummy;
+    avatar dummy;
 
     item &water = dummy.i_add( item( "water_clean", 0, 2 ) );
     item &bottle = dummy.i_add( item( "bottle_plastic" ) );
